recommender/requirements.txt
flask
python-dotenv
psycopg2-binary
scikit-learn
scipy
numpy